Move common to other dir
