+++ jni/jni_common/jni_bitmap2mat.h
