+++ jni/jni_common/jni_utils.h
