+++ jni/jni_common/types.h
