+++ jni/jni_common/yuv2rgb.h
@@ -8,7 +8,7 @@
  */
 
 #pragma once
-#include <common/types.h>
+#include <jni_common/types.h>
 
 namespace jnicommon {
