+++ jni/jni_detections/Android.mk
@@ -15,18 +15,14 @@ LOCAL_C_INCLUDES +=  \
 LOCAL_SRC_FILES += \
            jni_imageutils.cpp \
 	       jni_pedestrian_det.cpp \
-	       jni_face_det.cpp \
-           common/rgb2yuv.cpp \
-           common/yuv2rgb.cpp \
-           common/jni_bitmap2mat.cpp \
-           common/jni_utils.cpp \
-           common/jni_fileutils.cpp
+	       jni_face_det.cpp
 
 LOCAL_LDLIBS += -lm -llog -ldl -lz -ljnigraphics
 LOCAL_CPPFLAGS += -fexceptions -frtti -std=c++11
 
 # import dlib
-LOCAL_STATIC_LIBRARIES += dlib
+LOCAL_STATIC_LIBRARIES += dlib \
+                          jni_common
 
 ### import miniglog
 ifeq ($(MINIGLOG_LIB_TYPE),SHARED)
